Remove WP_Fields_API_Image_Control
This only exists in the customizer for back-compat. It contained empty deprecated functions related to a UI that is no longer used and adjusted the button labels to use "image" instead of "file", which isn't enough of a change to justify an entirely separate control. Use WP_Fields_API_Media_Control or Upload control with the `mime_type` argument set to `image` to achieve the same results.